+++ app/src/main/java/com/starters/hsge/domain/repository/ChatListRepository.kt
@@ -3,5 +3,5 @@ package com.starters.hsge.domain.repository
 import com.starters.hsge.presentation.main.chatroom.MessageInfo
 
 interface ChatListRepository {
-    suspend fun getChatList(roomId: Int): MessageInfo
+    suspend fun getChatList(roomId: Long): MessageInfo
 }
